pyproject.toml
[project]
name = "cal-reminders"
version = "0.2.0"
description = "Menubar countdown to your next calendar event"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rumps>=0.4.0",
    "pyobjc-framework-EventKit>=10.0",
]

[project.scripts]
cal-reminders = "cal_reminders.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cal_reminders"]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
